+++ uniffi_bindgen/src/macro_metadata/ci.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::interface::{ComponentInterface, Enum, Record, Type};
+use crate::interface::{CallbackInterface, ComponentInterface, Enum, Record, Type};
 use anyhow::{bail, Context};
 use uniffi_meta::{group_metadata, EnumMetadata, ErrorMetadata, Metadata, MetadataGroup};
 
@@ -96,6 +96,12 @@ fn add_item_to_ci(iface: &mut ComponentInterface, item: Metadata) -> anyhow::Res
         }
         Metadata::Object(meta) => {
             iface.add_object_free_fn(meta)?;
+        }
+        Metadata::CallbackInterface(meta) => {
+            iface.add_callback_interface_definition(CallbackInterface::new(meta.name));
+        }
+        Metadata::CallbackInterfaceMethod(meta) => {
+            iface.add_callback_interface_method_meta(meta)?;
         }
         Metadata::Error(meta) => {
             iface.note_name_used_as_error(meta.name());