+++ uniffi_macros/src/error.rs
@@ -71,14 +71,14 @@ fn error_ffi_converter_impl(ident: &Ident, enum_: &DataEnum, attr: &ErrorAttr) -
             enum_,
             attr.tag.as_ref(),
             attr.with_try_read.is_some(),
-            attr.callback_error.is_some(),
+            attr.handle_unknown_callback_error.is_some(),
         )
     } else {
         rich_error_ffi_converter_impl(
             ident,
             enum_,
             attr.tag.as_ref(),
-            attr.callback_error.is_some(),
+            attr.handle_unknown_callback_error.is_some(),
         )
     }
 }
@@ -92,7 +92,7 @@ fn flat_error_ffi_converter_impl(
     enum_: &DataEnum,
     tag: Option<&Path>,
     implement_try_read: bool,
-    callback_error: bool,
+    handle_unknown_callback_error: bool,
 ) -> TokenStream {
     let name = ident_to_string(ident);
     let impl_spec = tagged_impl_header("FfiConverter", ident, tag);
@@ -135,7 +135,8 @@ fn flat_error_ffi_converter_impl(
         quote! { ::std::panic!("try_read not supported for flat errors") }
     };
 
-    let handle_callback_unexpected_error = handle_callback_unexpected_error_fn(callback_error);
+    let handle_callback_unexpected_error =
+        handle_callback_unexpected_error_fn(handle_unknown_callback_error);
 
     quote! {
         #[automatically_derived]
@@ -197,7 +198,7 @@ mod kw {
     syn::custom_keyword!(tag);
     syn::custom_keyword!(flat_error);
     syn::custom_keyword!(with_try_read);
-    syn::custom_keyword!(callback_error);
+    syn::custom_keyword!(handle_unknown_callback_error);
 }
 
 #[derive(Default)]
@@ -206,7 +207,7 @@ pub(crate) struct ErrorAttr {
     flat: Option<kw::flat_error>,
     with_try_read: Option<kw::with_try_read>,
     /// Can this error be used in a callback interface?
-    callback_error: Option<kw::callback_error>,
+    handle_unknown_callback_error: Option<kw::handle_unknown_callback_error>,
 }
 
 impl UniffiAttributeArgs for ErrorAttr {
@@ -229,9 +230,9 @@ impl UniffiAttributeArgs for ErrorAttr {
                 with_try_read: input.parse()?,
                 ..Self::default()
             })
-        } else if lookahead.peek(kw::callback_error) {
+        } else if lookahead.peek(kw::handle_unknown_callback_error) {
             Ok(Self {
-                callback_error: input.parse()?,
+                handle_unknown_callback_error: input.parse()?,
                 ..Self::default()
             })
         } else {
@@ -244,7 +245,10 @@ impl UniffiAttributeArgs for ErrorAttr {
             tag: either_attribute_arg(self.tag, other.tag)?,
             flat: either_attribute_arg(self.flat, other.flat)?,
             with_try_read: either_attribute_arg(self.with_try_read, other.with_try_read)?,
-            callback_error: either_attribute_arg(self.callback_error, other.callback_error)?,
+            handle_unknown_callback_error: either_attribute_arg(
+                self.handle_unknown_callback_error,
+                other.handle_unknown_callback_error,
+            )?,
         })
     }
 }